+++ liquid/conf/forms.py
@@ -1,5 +1,5 @@
 from django.forms import ModelForm
-from conference.models import Company
+from conf.models import Company
 
 class CompanyForm(ModelForm):
     class Meta:
