--- liquid/conference/urls.py
@@ -1,6 +1,0 @@
-from django.conf.urls.defaults import patterns, include, url
-
-urlpatterns = patterns('',
-    # Examples:
-    url(r'^2014/jobfair$', 'conference.views.jobfair'),
-)
