+++ liquid/templates/conf/base.html
@@ -1,0 +1,83 @@
+<!DOCTYPE html>
+<html lang="en">
+  <head>
+    <meta charset="utf-8">
+    <title>ACM Reflections | Projections 2013</title>
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+
+    <!-- Le styles -->
+	<link href='http://fonts.googleapis.com/css?family=Dancing+Script:700' rel='stylesheet' type='text/css'>
+    <link href="/static/css/bootstrap.min.css" rel="stylesheet">
+    <link href="/static/css/conference.css" rel="stylesheet">
+
+    {% block top_styles %}
+    {% endblock %}
+
+    <!-- Le HTML5 shim, for IE6-8 support of HTML5 elements -->
+    <!--[if lt IE 9]>
+      <script src="http://html5shim.googlecode.com/svn/trunk/html5.js"></script>
+    <![endif]-->
+
+    <!-- Le fav and touch icons -->
+    <link rel="shortcut icon" href="/static/favicon.ico">
+    <script type="text/javascript">
+
+       var _gaq = _gaq || [];
+       _gaq.push(['_setAccount', 'UA-30929649-1']);
+       _gaq.push(['_trackPageview']);
+
+       (function() {
+         var ga = document.createElement('script'); ga.type = 'text/javascript'; ga.async = true;
+         ga.src = ('https:' == document.location.protocol ? 'https://ssl' : 'http://www') + '.google-analytics.com/ga.js';
+         var s = document.getElementsByTagName('script')[0]; s.parentNode.insertBefore(ga, s);
+       })();
+
+     </script>
+  </head>
+  <body>
+    <div id="the-meat" class="container">
+      <div class="container">
+        <div class="header">
+          <!--<div id="triangle-bar"></div>-->
+          <!-- <ul class="navbar-fixed-bottom nav-pills main_nav">
+            <li{% if section == 'about' %} class='active'{% endif %}><a href="/about">About</a></li>
+            <li{% if section == 'calendar' %} class='active'{% endif %}><a href="/calendar">Calendar</a></li>
+            <li{% if section == 'sigs' %} class='active'{% endif %}><a href="/sigs">SIGs</a></li>
+            <li{% if section == 'conference' %} class='active'{% endif %}><a href="/conf">Conference</a></li>
+            <li{% if section == 'corporate' %} class='active'{% endif %}><a href="/corporate">Corporate</a></li>
+            <li{% if section == 'contact' %} class='active'{% endif %}><a href="/contact">Contact</a></li>
+            <li{% if section == 'intranet' %} class='active'{% endif %}><a href="/intranet">Intranet</a></li>
+            <ul class="nav nav-pills pull-right right_nav">
+            </ul>
+          </ul> -->
+          
+        </div>
+      </div>
+      
+
+      <div class="container main">
+          {% block content %}
+          {% endblock %}
+      </div> <!-- /container -->
+      
+      <footer class="container">
+      <!--ACM@ILLINOIS | <a href="tel:1-217-333-5828">(217)-333-5828</a> | <a href="mailto:acm@illinois.edu">acm@illinois.edu</a> | Ste. 1104 Thomas M. Siebel Center for Computer Science, 201 N. Goodwin Ave. Urbana, IL 61801-->
+      </footer>
+    </div>
+
+    <!-- Le javascript
+    ================================================== -->
+    <!-- Placed at the end of the document so the pages load faster -->
+    <script src="/static/js/jquery-1.8.3.min.js"></script>
+    <!-- <script src="/static/js/jquery.tiles-gallery.js"></script>
+    <script type="text/javascript">
+      $(function () {
+        $(".tiles").tilesGallery({
+          height: 500,
+        });
+      })
+    </script>-->
+    {% block bottom_scripts %}
+    {% endblock %}
+  </body>
+</html>
