+++ liquid/templates/conf/job-fair-invite.html
@@ -1,0 +1,36 @@
+{% extends "conf/base.html" %}
+{% load bootstrap %}
+{% block content %}
+<div class="row">
+  <div class="span12">
+	<div id="job-fair-form">
+        <img class="header-logo" src="/static/img/conference/header-logo.png" />
+        {% if company.type == JOBFAIR %}
+            <a name="form1570856351" id="formAnchor1570856351"></a>
+            <script type="text/javascript" src="http://fs6.formsite.com/include/form/embedManager.js?1570856351"></script>
+            <script type="text/javascript">
+                EmbedManager.embed({
+                    key: "http://fs6.formsite.com/res/showFormEmbed?EParam=0f3EXAbWkRyOWvWrn2%2FGghGOwQgdGLHi&1570856351",
+                    width: "100%"
+                });
+            </script>
+        {% elif company.type == STARTUP %}
+            <a name="form1613832842" id="formAnchor1613832842"></a>
+            <script type="text/javascript" src="http://fs6.formsite.com/include/form/embedManager.js?1613832842"></script>
+            <script type="text/javascript">
+                EmbedManager.embed({
+                    key: "http://fs6.formsite.com/res/showFormEmbed?EParam=0f3EXAbWkRyOWvWrn2%2FGghjGupL6arfd&1613832842",
+                    width: "100%"
+                });
+            </script>
+        {% endif %}
+    </div>
+  </div>
+</div>
+{% endblock %}
+
+{% block bottom_scripts %}
+<script type="text/javascript">
+  $("#the-meat").height("1500");
+</script>
+{% endblock %}
