+++ liquid/urls.py
@@ -19,7 +19,7 @@ urlpatterns = patterns('',
     url(r'^login/$', 'django.contrib.auth.views.login'),
     url(r'^logout/$', 'django.contrib.auth.views.logout'),
     url(r'^intranet/', include('intranet.urls')),
-    url(r'^conference/', include('conference.urls')),
+    #url(r'^conference/', include('conference.urls')),
     url(r'^kiosk/', 'kiosk.views.main'),
 
     url(r'^resume/$', redirect_to, {'url': '/corporate/resume/'}),
@@ -27,15 +27,15 @@ urlpatterns = patterns('',
         'https://www-s.acm.uiuc.edu/confluence/display/admin/Printing'}),
 
     #rp redirects
-    url(r'^mm/$',redirect_to, {'url': '/conference/2013/mechmania.html'}),
-    url(r'^attend/$',redirect_to, {'url': '/conference/2013/register.html'}),
-    url(r'^register/$',redirect_to, {'url': '/conference/2013/register.html'}),
-    url(r'^helpout/$',redirect_to, {'url': '/conference/2013/volunteers.html'}),
-    url(r'^schedule/$',redirect_to, {'url': '/conference/2013/schedule.html'}),
-    url(r'^jobs/$',redirect_to, {'url': '/conference/2013/career_fairs.html'}),
-    url(r'^jobfair/$',redirect_to, {'url': '/conference/2013/career_fairs.html'}),
-    url(r'^startup/$',redirect_to, {'url': '/conference/2013/career_fairs.html'}),
-    url(r'^startupfair/$',redirect_to, {'url': '/conference/2013/career_fairs.html'}),
+    url(r'^mm/$',redirect_to, {'url': '/conference/2014/mechmania.html'}),
+    url(r'^attend/$',redirect_to, {'url': '/conference/2014/register.html'}),
+    url(r'^register/$',redirect_to, {'url': '/conference/2014/register.html'}),
+    url(r'^helpout/$',redirect_to, {'url': '/conference/2014/volunteers.html'}),
+    url(r'^schedule/$',redirect_to, {'url': '/conference/2014/schedule.html'}),
+    url(r'^jobs/$',redirect_to, {'url': '/conference/2014/career_fairs.html'}),
+    url(r'^jobfair/$',redirect_to, {'url': '/conference/2014/career_fairs.html'}),
+    url(r'^startup/$',redirect_to, {'url': '/conference/2014/career_fairs.html'}),
+    url(r'^startupfair/$',redirect_to, {'url': '/conference/2014/career_fairs.html'}),
     url(r'^mm-irc/$',redirect_to, {'url': 'http://webchat.freenode.net/?channels=%23%23mm19&uio=OT10cnVlde'}),
 
     #sigmusic redirects
